cs.cmdutils: pipefrom: new `keep_stdin` mode
--HG--
branch : pypi

+++ lib/python/cs/cmdutils.py
@@ -23,7 +23,7 @@ def run(argv, trace=False):
     print(*pargv, file=tracefp)
   return subprocess.call(argv)
 
-def pipefrom(argv, trace=False, binary=False, **kw):
+def pipefrom(argv, trace=False, binary=False, keep_stdin=False, **kw):
   ''' Pipe text from a command. Optionally trace invocation. Return the Popen object with .stdout decoded as text.
       `argv`: the command argument list
       `binary`: if true (default false) return the binary stdout instead of a text wrapper
@@ -36,18 +36,21 @@ def pipefrom(argv, trace=False, binary=False, **kw):
     tracefp = sys.stderr if trace is True else trace
     pargv = ['+'] + list(argv) + ['|']
     print(*pargv, file=tracefp)
-  sp_devnull = getattr(subprocess, 'DEVNULL', None)
-  if sp_devnull is None:
-    devnull = open(os.devnull, 'wb')
-  else:
-    devnull = sp_devnull
-  P = subprocess.Popen(argv, stdin=devnull, stdout=subprocess.PIPE)
+  popen_kw = {}
+  if not keep_stdin:
+    sp_devnull = getattr(subprocess, 'DEVNULL', None)
+    if sp_devnull is None:
+      devnull = open(os.devnull, 'wb')
+    else:
+      devnull = sp_devnull
+    popen_kw['stdin'] = devnull
+  P = subprocess.Popen(argv, stdout=subprocess.PIPE, **popen_kw)
   if binary:
     if kw:
       raise ValueError("binary mode: extra keyword arguments not supported: %r", kw)
   else:
     P.stdout = io.TextIOWrapper(P.stdout, **kw)
-  if sp_devnull is None:
+  if not keep_stdin and sp_devnull is None:
     devnull.close()
   return P
